Applied additional patch for HARMONY-3725 ([drlvm][verifier] Eclipse help doesn't work) and patch that fixes bug in HARMONY-2281 [drlvm][verifier] Remove assignment compatible check for aastore instruction
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@532049 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/vmcore/src/verifier/ver_subroutine.cpp
@@ -142,7 +142,7 @@ static inline vf_Result ResolveSubroutineEntryPoint(vf_Node *node,
         }
         map->m_stack[index].m_type = SM_RETURN_ADDR;
         map->m_stack[index].m_pc =
-            p_element->m_node->m_start->m_addr - ctx->m_bytes;
+            (unsigned) (p_element->m_node->m_start->m_addr - ctx->m_bytes);
     }
 
     VF_DUMP(DUMP_NODESTACK, DumpNodeStack(ctx));
